backend/requirement.txt
flask
flask-cors
psycopg2
bcrypt
google-generativeai
